remove Record.java move TsFileHeader from TsFile to ConcurrentBuffer

+++ storage-module/src/main/java/org/luncert/tinystorage/storemodule/descriptor/TsBucketDesc.java
@@ -17,7 +17,7 @@ public class TsBucketDesc extends Descriptor {
 
   private String id;
 
-  private int bucketSize;
+  private long bucketSize;
 
   private long lastAccessTime;
